requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
Pillow==10.1.0
gunicorn==21.2.0
cloudinary==1.42.2
